Add files via upload
Exploratory Data Analysis (EDA) is a crucial step in the data science workflow. It involves summarizing the main characteristics of a dataset, often with visual methods, before applying any machine learning models. The goal of EDA is to understand the data's structure, detect anomalies, test hypotheses, and check assumptions with the help of summary statistics and graphical representations. By doing so, data scientists can identify patterns, spot outliers, and gain insights that guide the development of more precise and effective models. EDA is both an art and a science, requiring a keen eye and analytical skills.